TrendingItems service integrated to Network

+++ TheMovieDB/NetworkModels/Models/MovieDetail.swift
@@ -9,7 +9,7 @@ import Foundation
 
 public struct MovieDetail: Decodable {
     var id: Int?
-    var voteAverage: Int?
+    var voteAverage: Double?
     var originalTitle: String?
     var overview: String?
     var posterPath: String?
@@ -18,7 +18,7 @@ public struct MovieDetail: Decodable {
     
     public init(from decoder: Decoder) throws {
         let values = try decoder.container(keyedBy: CodingKeys.self)
-        voteAverage = try values.decode(Int.self, forKey: .voteAverage)
+        voteAverage = try values.decode(Double.self, forKey: .voteAverage)
         originalTitle = try values.decode(String.self, forKey: .originalTitle)
         posterPath = try values.decode(String.self, forKey: .posterPath)
         backdropPath = try values.decode(String.self, forKey: .backdropPath)
